+++ src/components/tasks/task-detail.jsx
@@ -40,7 +40,7 @@ const TaskDetail = () => {
   const filteredTasks = tasks.filter((task) => task.projectId === projectId);
 
   const formatDate = (isoString) => {
-    const options = { year: 'numeric', month: 'long', day: 'numeric' };
+    const options = { year: 'numeric', month: 'numeric', day: 'numeric' };
     return new Date(isoString).toLocaleDateString(undefined, options);
   };
